+++ server/src/main/java/seb39_40/coffeewithme/review/dto/ReviewResponseDto.java
@@ -4,11 +4,7 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import seb39_40.coffeewithme.review.domain.ReviewTag;
-import seb39_40.coffeewithme.user.dto.UserResponseDto;
-
-import java.util.List;
-import java.util.stream.Collectors;
+import seb39_40.coffeewithme.user.dto.response.UserResponseDto;
 
 @Data @NoArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
