+++ server/src/main/java/seb39_40/coffeewithme/review/mapper/ReviewMapper.java
@@ -6,16 +6,14 @@ import seb39_40.coffeewithme.image.domain.Image;
 import seb39_40.coffeewithme.review.domain.Review;
 import seb39_40.coffeewithme.review.dto.ReviewRequestDto;
 import seb39_40.coffeewithme.review.dto.ReviewResponseDto;
-import seb39_40.coffeewithme.tag.domain.Tag;
 import seb39_40.coffeewithme.user.mapper.UserMapper;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.LinkedHashMap;
 import java.util.List;
 
 import static seb39_40.coffeewithme.review.dto.ReviewResponseDto.*;
-import static seb39_40.coffeewithme.user.dto.UserResponseDto.*;
+import static seb39_40.coffeewithme.user.dto.response.UserResponseDto.*;
 
 @Mapper(componentModel = "spring", uses = {UserMapper.class, Collections.class})
 public interface ReviewMapper {
