+++ server/src/main/java/seb39_40/coffeewithme/wishlist/dto/WishlistResponse.java
@@ -1,4 +1,4 @@
-package seb39_40.coffeewithme.user.dto;
+package seb39_40.coffeewithme.wishlist.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
